+++ source/tools/atlas/GameInterface/Handlers/ObjectHandlers.cpp
@@ -316,7 +316,7 @@ QUERYHANDLER(SelectObject)
 	msg->pos.GetScreenSpace(x, y);
 	
 	CVector3D rayorigin, raydir;
-	g_Game->GetView()->GetCamera()->BuildCameraRay(x, y, rayorigin, raydir);
+	g_Game->GetView()->GetCamera()->BuildCameraRay((int)x, (int)y, rayorigin, raydir);
 
 	CUnit* target = g_UnitMan.PickUnit(rayorigin, raydir);
